Correct faiing tests - WIP
