+++ .grenrc.yml
@@ -1,0 +1,6 @@
+---
+  dataSource: "issues"
+  #prefix: "v"
+  onlyMilestones: false
+  groupBy: "label"
+  changelogFilename: "CHANGELOG.md"
