+++ src/test/groovy/com/github/prokod/gradle/crossbuild/CrossBuildPluginPublishingTest.groovy
@@ -23,7 +23,7 @@ import spock.lang.Unroll
 
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
-class CrossBuildPluginImplementationMultiModuleTest extends CrossBuildGradleRunnerSpec {
+class CrossBuildPluginPublishingTest extends CrossBuildGradleRunnerSpec {
 
     File settingsFile
     File propsFile
@@ -54,19 +54,13 @@ class CrossBuildPluginImplementationMultiModuleTest extends CrossBuildGradleRunn
     }
 
     /**
-     * Test Properties:
-     * <ul>
-     *     <li>Plugin apply mode: Eager</li>
-     * </ul>
-     * This test is using 'implementation' configuration to express dependency graph between the modules.
-     *
      * resource file/s for the test:
      * 04-app_builds_resolved_configurations.json
      * 04-pom_lib2-00.xml
      * 04-pom_app-00.xml
      */
     @Unroll
-    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin on a multi-module project with dependency graph of depth 3 and with cross building dsl that is different on each submodule and with publishing dsl, should produce expected jars and pom files and should have correct pom files content"() {
+    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin on a multi-module project with dependency graph of depth 3 and with cross building dsl that is different on each submodule and with publishing dsl, should produce expected jars and pom files and should have correct pom files content blah"() {
         given:
         // root project settings.gradle
         settingsFile << """
@@ -103,6 +97,10 @@ subprojects {
     }
     
     if (!project.name.endsWith('app')) {
+        java {
+            withSourcesJar()
+            withJavadocJar()
+        }
         publishing {
             publications {
                 crossBuildSpark230_211(MavenPublication) {
@@ -355,256 +353,8 @@ dependencies {
 
         where:
         gradleVersion | defaultScalaVersion
-        '4.10.3'      | '2.12'
-        '5.6.4'       | '2.11'
+        // '4.10.3'      | '2.12'
+        // '5.6.4'       | '2.11'
         '6.5'         | '2.12'
-        '7.2'         | '2.11'
-    }
-
-    /**
-     * Test Properties:
-     * <ul>
-     *     <li>Plugin apply mode: Lazy</li>
-     * </ul>
-     * Here lib3 is a non cross build dependency.
-     * This test checks that when lib3 is added to dependent app module, while using 'implementation' configuration,
-     * cross build compileScala scenario should not fail because of a missing lib3 project dependency in
-     * compileClasspath
-     *
-     * @return
-     */
-    // todo add pom checks
-    @Unroll
-    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin on a multi-module project with dependency graph of depth 3 and with cross building dsl that is applied only to some sub modules should produce expected: jars, pom files; and pom files content should be correct"() {
-        given:
-        // root project settings.gradle
-        settingsFile << """
-include 'lib'
-include 'lib3'
-include 'app'
-"""
-
-        buildFile << """
-plugins {
-    id 'com.github.prokod.gradle-crossbuild' apply false
-}
-
-allprojects {
-    group = 'com.github.prokod.it'
-    version = '1.0-SNAPSHOT'
-    
-    repositories {
-        mavenCentral()
-    }
-}
-
-subprojects {
-    project.pluginManager.withPlugin('com.github.prokod.gradle-crossbuild') {
-        if (!project.name.endsWith('app')) {
-            crossBuild {
-                builds {
-                    spark233_211
-                    spark242_212
-                    spark243 {
-                        scalaVersions = ['2.11', '2.12']
-                        archive.appendixPattern = '-2-4-3_?'
-                    }
-                }
-            }
-        }
-    }
-
-    project.pluginManager.withPlugin('maven-publish') {
-        if (!project.name.endsWith('app')) {
-            publishing {
-                publications {
-                    crossBuildSpark233_211(MavenPublication) {
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
-                            artifact crossBuildSpark233_211Jar
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
-                    }
-                    crossBuildSpark242_212(MavenPublication) {
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
-                            artifact crossBuildSpark242_212Jar
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
-                    }
-                    crossBuildSpark243_211(MavenPublication) {
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
-                            artifact crossBuildSpark243_211Jar
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
-                    }
-                    crossBuildSpark243_212(MavenPublication) {
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
-                            artifact crossBuildSpark243_212Jar
-                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
-                    }
-                }
-            }
-        }
-    }
-
-    tasks.withType(GenerateMavenPom) { t ->
-        if (t.name.contains('CrossBuildSpark233_211')) {
-            t.destination = file("\$buildDir/generated-pom233_2.11.xml")
-        }
-        if (t.name.contains('CrossBuildSpark242_212')) {
-            t.destination = file("\$buildDir/generated-pom242_2.12.xml")
-        }
-    }
-}
-"""
-
-        libScalaFile << """
-
-object scalazOption {
-  import scalaz._
-  import Scalaz._ 
-  val boolT = 6 < 10
-  
-  boolT.option("corrie")
-}
-
-object Factorial {
-  // The actual implementation is regular old-fashioned scala code:
-  def normalFactorial(n: Int): Int =
-    if (n == 0) 1
-    else n * normalFactorial(n - 1)
-}
-"""
-
-        libBuildFile << """
-apply plugin: 'com.github.prokod.gradle-crossbuild'
-apply plugin: 'maven-publish'
-
-crossBuild {
-    builds {
-        v210 {
-            archive.appendixPattern = '-legacy_?'
-        }
-    }
-}
-
-sourceSets {
-    main {
-        scala {
-            srcDirs = ['src/main/scala', 'src/main/java']
-        }
-        java {
-            srcDirs = []
-        }
-    }
-}
-
-dependencies {
-    implementation 'org.scalaz:scalaz-core_${defaultScalaVersion}:7.2.28'
-    crossBuildV210Implementation 'org.scalaz:scalaz-core_2.10:7.2.25'
-    crossBuildSpark233_211Implementation 'org.scalaz:scalaz-core_2.11:7.2.26'
-    crossBuildSpark242_212Implementation 'org.scalaz:scalaz-core_2.12:7.2.27'
-    crossBuildSpark243_211Implementation 'org.scalaz:scalaz-core_2.11:7.2.28'
-    crossBuildSpark243_212Implementation 'org.scalaz:scalaz-core_2.12:7.2.28'
-
-    implementation 'org.scala-lang:scala-library:${defaultScalaVersion}.+'
-}
-"""
-
-        lib3JavaFile << """
-import org.apache.commons.math3.util.MathUtils;
-
-public class CompileTimeFactorialUtils {
-
-   public static String printHello() {
-      double twoPi = MathUtils.TWO_PI;
-      return "Hello! " + twoPi;
-   }
-}
-"""
-
-        lib3BuildFile << """
-apply plugin: 'java'
-//apply plugin: 'maven-publish'
-
-dependencies {
-  implementation 'org.apache.commons:commons-math3:3.6.1'
-}
-"""
-
-        appScalaFile << """
-object Test extends App {
-    import Factorial._
-    import CompileTimeFactorialUtils._
-
-    println(normalFactorial(10))
-    println(CompileTimeFactorialUtils.printHello)
-}
-"""
-
-        appBuildFile << """
-apply plugin: 'com.github.prokod.gradle-crossbuild'
-apply plugin: 'maven-publish'
-
-crossBuild {
-    builds {
-        spark233_211 {
-            archive.appendixPattern = '-all_?'
-        }
-    }
-}
-
-publishing {
-    publications {
-        crossBuildSpark233_211(MavenPublication) {
-            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
-                artifact crossBuildSpark233_211Jar
-            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
-        }
-    }
-}
-        
-dependencies {
-    implementation project(':lib')
-    implementation project(':lib3')
-
-    implementation 'org.scala-lang:scala-library:${defaultScalaVersion}.+'
-}
-"""
-
-        when:
-        Assume.assumeTrue(testMavenCentralAccess())
-        def result = GradleRunner.create()
-                .withGradleVersion(gradleVersion)
-                .withProjectDir(dir.root)
-                .withPluginClasspath()
-                .withDebug(true)
-                .withArguments('build', 'lib:crossBuildV210Jar', 'lib3:jar', 'app:crossBuildSpark233_211Jar', 'app:crossBuildResolvedConfigs', '--info', '--stacktrace')
-                .build()
-
-        then:
-        result.task(":lib:build").outcome == SUCCESS
-        result.task(":lib3:build").outcome == SUCCESS
-        result.task(":app:build").outcome == SUCCESS
-        result.task(":app:crossBuildResolvedConfigs").outcome == SUCCESS
-        result.task(":lib3:jar").outcome == SUCCESS
-        result.task(":lib:crossBuildV210Jar").outcome == SUCCESS
-        result.task(":lib:crossBuildSpark233_211Jar").outcome == SUCCESS
-        result.task(":app:crossBuildSpark233_211Jar").outcome == SUCCESS
-
-        // 'build' task should:
-        fileExists("$dir.root.absolutePath/lib/build/libs/lib-1.0-SNAPSHOT.jar")
-        fileExists("$dir.root.absolutePath/lib3/build/libs/lib3-1.0-SNAPSHOT.jar")
-        fileExists("$dir.root.absolutePath/app/build/libs/app-1.0-SNAPSHOT.jar")
-
-        // 'lib:crossBuildV210Jar', 'app:crossBuildSpark233_211Jar' tasks should:
-        fileExists("$dir.root.absolutePath/lib/build/libs/lib-legacy_2.10*.jar")
-        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.11*.jar")
-        !fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.12*.jar")
-        fileExists("$dir.root.absolutePath/app/build/libs/app-all_2.11*.jar")
-        !fileExists("$dir.root.absolutePath/app/build/libs/app-all_2.12*.jar")
-
-        where:
-        gradleVersion   | defaultScalaVersion
-        '4.10.3'        | '2.11'
-        '5.6.4'         | '2.12'
-        '6.5'           | '2.12'
-        '7.2'           | '2.11'
     }
 }
